+++ src/main/java/com/c3/C3Application.java
@@ -1,4 +1,4 @@
-package c3;
+package com.c3;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
