+++ src/main/java/com/c3/login/entity/User.java
@@ -1,4 +1,4 @@
-package c3.login.entity;
+package com.c3.login.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
